fix : 머지하면서 생긴 문제들 갈아엎기
- Integer -> Long

+++ src/main/java/keeper/project/homepage/dto/PostingDto.java
@@ -33,7 +33,7 @@ public class PostingDto {
   private Integer isNotice;
   private Integer isSecret;
   private String password;
-  private Integer thumbnailId;
+  private Long thumbnailId;
 
   public PostingEntity toEntity(CategoryEntity categoryEntity, MemberEntity memberEntity,
       ThumbnailEntity thumbnailEntity) {
