+++ src/main/java/keeper/project/homepage/repository/OriginalImageRepository.java
@@ -3,6 +3,6 @@ package keeper.project.homepage.repository;
 import keeper.project.homepage.entity.OriginalImageEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OriginalImageRepository extends JpaRepository<OriginalImageEntity, Integer> {
+public interface OriginalImageRepository extends JpaRepository<OriginalImageEntity, Long> {
 
 }
