+++ src/main/java/keeper/project/homepage/repository/ThumbnailRepository.java
@@ -3,6 +3,6 @@ package keeper.project.homepage.repository;
 import keeper.project.homepage.entity.ThumbnailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ThumbnailRepository extends JpaRepository<ThumbnailEntity, Integer> {
+public interface ThumbnailRepository extends JpaRepository<ThumbnailEntity, Long> {
 
 }
